Move entities to model package
