+++ cinema/src/main/java/com/drbsimon/cinema/data_sample/RoomCreator.java
@@ -1,6 +1,6 @@
 package com.drbsimon.cinema.data_sample;
 
-import com.drbsimon.cinema.entity.Room;
+import com.drbsimon.cinema.model.Room;
 import com.drbsimon.cinema.repository.RoomRepository;
 import com.drbsimon.cinema.repository.SeatRepository;
 import org.springframework.stereotype.Component;
