+++ cinema/src/main/java/com/drbsimon/cinema/repository/RoomManager.java
@@ -1,7 +1,7 @@
 package com.drbsimon.cinema.repository;
 
-import com.drbsimon.cinema.entity.Room;
-import com.drbsimon.cinema.model.RoomListWrapper;
+import com.drbsimon.cinema.model.Room;
+import com.drbsimon.cinema.model.wrapper.RoomListWrapper;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
