+++ cinema/src/test/java/com/drbsimon/cinema/SeatTest.java
@@ -1,8 +1,8 @@
 package com.drbsimon.cinema;
 
-import com.drbsimon.cinema.entity.Room;
-import com.drbsimon.cinema.entity.Seat;
-import com.drbsimon.cinema.model.SeatListWrapper;
+import com.drbsimon.cinema.model.Room;
+import com.drbsimon.cinema.model.Seat;
+import com.drbsimon.cinema.model.wrapper.SeatListWrapper;
 import com.drbsimon.cinema.repository.RoomRepository;
 import com.drbsimon.cinema.repository.SeatManager;
 import com.drbsimon.cinema.repository.SeatRepository;
