+++ cinema/src/test/java/com/drbsimon/cinema/controller/RoomControllerTest.java
@@ -1,7 +1,7 @@
 package com.drbsimon.cinema.controller;
 
-import com.drbsimon.cinema.entity.Room;
-import com.drbsimon.cinema.model.RoomListWrapper;
+import com.drbsimon.cinema.model.Room;
+import com.drbsimon.cinema.model.wrapper.RoomListWrapper;
 import com.drbsimon.cinema.repository.RoomManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
