+++ cinema/src/test/java/com/drbsimon/cinema/controller/SeatControllerTest.java
@@ -1,8 +1,8 @@
 package com.drbsimon.cinema.controller;
 
-import com.drbsimon.cinema.entity.Room;
-import com.drbsimon.cinema.entity.Seat;
-import com.drbsimon.cinema.model.SeatListWrapper;
+import com.drbsimon.cinema.model.Room;
+import com.drbsimon.cinema.model.Seat;
+import com.drbsimon.cinema.model.wrapper.SeatListWrapper;
 import com.drbsimon.cinema.repository.SeatManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
